uploader/requirements.txt
minio
Pillow
requests